+++ src/date.ts
@@ -3,7 +3,7 @@ import isSameOrAfter from 'dayjs/plugin/isSameOrAfter.js';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore.js';
 import timezone from 'dayjs/plugin/timezone.js';
 import utc from 'dayjs/plugin/utc.js';
-import { get } from 'lodash';
+import { get } from 'lodash-es';
 import { DateOperator } from './operator';
 import type { DateInputValue, DateRule } from './types';